+++ pkg/hashlib/common.go
