+++ pkg/hashlib/common_test.go
