+++ pkg/hashlib/hasher.go
